+++ recipes/client.rb
@@ -1,0 +1,2 @@
+node.default['consul']['service_mode'] = 'client'
+include_recipe 'consul'