+++ recipes/cluster.rb
@@ -1,0 +1,2 @@
+node.default['consul']['service_mode'] = 'cluster'
+include_recipe 'consul'