+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.2.0
+
+* adds ubuntu support
+* adds cluster & client recipes
+* adds serverspec tests
+* bumps consul version to 0.5.2
+* adds monit as consul watchdog
+
 # 0.1.0
 
 Initial release of consul